gateway y workers andando con fallas

+++ utils/DatasetHandler.py
@@ -1,3 +1,4 @@
+import os
 from unittest import TestCase
 import csv
 from utils.auxiliar_functions import integer_to_big_endian_byte_array, byte_array_to_big_endian_integer, remove_bytes, recv_exactly
@@ -75,11 +76,14 @@ class DatasetReader():
         self.file.close()
 
 class DatasetWriter():
-    def __init__(self, path, columns):
+    def __init__(self, path, columns, append=False):
         self.file = open(path, 'a', encoding='Utf-8')
-        self.file.seek(0)
-        self.file.truncate()
-        self.header = False
+        print("append was ", append)
+        if not append:
+            self.file.seek(0)
+            self.file.truncate()
+            self.file.flush()
+            os.fsync(self.file.fileno())
         self.writer = csv.DictWriter(self.file, fieldnames=columns, lineterminator='\n')
 
     def append_objects(self, objects):
@@ -95,9 +99,8 @@ class DatasetWriter():
         Appends an object using the get_csv_values method, which must ruturn
         the amount of values specified when the Writer was created
         """
-        if not self.header:
+        if self.file.tell() == 0:
             self.writer.writeheader()
-            self.header = True
         columns = self.writer.fieldnames
         values = object.get_csv_values()
         line = {}
@@ -160,6 +163,24 @@ if __name__ == '__main__':
             result1 = QueryMessage(BOOK_MSG_TYPE, title="Murdoca", authors="['Mazzeo']")
             result2 = QueryMessage(BOOK_MSG_TYPE, title="Fisica", authors="['Sears', 'Semanski']")
             dw.append_objects([result1, result2])
+            dw.close()
+
+            file = open('test_result.csv', 'r', encoding='Utf-8',)
+            reader = csv.DictReader(file)
+            values = [list(next(reader).values()), list(next(reader).values())]
+            file.close()
+            self.assertEqual(values, [result1.get_csv_values(), result2.get_csv_values()])
+
+        def test_append_on_existing_file(self):
+            columns = ["Title", "author"]
+            dw = DatasetWriter('test_result.csv', columns)
+            result1 = QueryMessage(BOOK_MSG_TYPE, title="Murdoca", authors="['Mazzeo']")
+            dw.append_objects([result1])
+            dw.close()
+
+            dw = DatasetWriter('test_result.csv', columns, append=True)
+            result2 = QueryMessage(BOOK_MSG_TYPE, title="Fisica", authors="['Sears', 'Semanski']")
+            dw.append_objects([result2])
             dw.close()
 
             file = open('test_result.csv', 'r', encoding='Utf-8',)
